+++ GamesToGo.Desktop/GamesToGo.Desktop.csproj
@@ -18,4 +18,8 @@
     <PackageReference Include="System.Data.SQLite.EF6" Version="1.0.112" />
   </ItemGroup>
 
+  <ItemGroup>
+    <Folder Include="Proyect\" />
+  </ItemGroup>
+
 </Project>
